PRJ: Corrected spelling of coalescence in file and class names
