+++ applications/solvers/multiphase/interfacialModels/Make/files
@@ -59,18 +59,18 @@ aspectRatioModels/Wellek/Wellek.C
 
 wallDependentModel/wallDependentModel.C
 
-coalesenceKernels/coalesenceKernel.C
+coalescenceKernels/coalescenceKernel.C
 
-coalesenceKernels/coalesenceFrequencyKernels/coalesenceFrequencyKernel/coalesenceFrequencyKernel.C
-coalesenceKernels/coalesenceFrequencyKernels/coalesenceFrequencyKernel/newCoalesenceFrequencyKernel.C
-coalesenceKernels/coalesenceFrequencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.C
-coalesenceKernels/coalesenceFrequencyKernels/PrinceAndBlanch/PrinceAndBlanch.C
+coalescenceKernels/coalescenceFrequencyKernels/coalescenceFrequencyKernel/coalescenceFrequencyKernel.C
+coalescenceKernels/coalescenceFrequencyKernels/coalescenceFrequencyKernel/newCoalescenceFrequencyKernel.C
+coalescenceKernels/coalescenceFrequencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.C
+coalescenceKernels/coalescenceFrequencyKernels/PrinceAndBlanch/PrinceAndBlanch.C
 
-coalesenceKernels/coalesenceEfficiencyKernels/coalesenceEfficiencyKernel/coalesenceEfficiencyKernel.C
-coalesenceKernels/coalesenceEfficiencyKernels/coalesenceEfficiencyKernel/newCoalesenceEfficiencyKernel.C
-coalesenceKernels/coalesenceEfficiencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.C
-coalesenceKernels/coalesenceEfficiencyKernels/Chesters/Chesters.C
-coalesenceKernels/coalesenceEfficiencyKernels/constant/constant.C
+coalescenceKernels/coalescenceEfficiencyKernels/coalescenceEfficiencyKernel/coalescenceEfficiencyKernel.C
+coalescenceKernels/coalescenceEfficiencyKernels/coalescenceEfficiencyKernel/newCoalescenceEfficiencyKernel.C
+coalescenceKernels/coalescenceEfficiencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.C
+coalescenceKernels/coalescenceEfficiencyKernels/Chesters/Chesters.C
+coalescenceKernels/coalescenceEfficiencyKernels/constant/constant.C
 
 bubbleBreakupKernels/bubbleBreakupKernel/bubbleBreakupKernel.C
 bubbleBreakupKernels/bubbleBreakupKernel/newBubbleBreakupKernel.C
