+++ applications/solvers/multiphase/interfacialModels/coalescenceKernels/coalescenceEfficiencyKernels/Chesters/Chesters.C
@@ -23,22 +23,23 @@ License
 
 \*---------------------------------------------------------------------------*/
 
-#include "CoulaloglouAndTavlarides.H"
+#include "Chesters.H"
 #include "addToRunTimeSelectionTable.H"
 #include "fundamentalConstants.H"
+#include "phasePair.H"
 
 // * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
 
 namespace Foam
 {
-namespace coalesenceFrequencyKernels
+namespace coalescenceEfficiencyKernels
 {
-    defineTypeNameAndDebug(CoulaloglouAndTavlarides, 0);
+    defineTypeNameAndDebug(Chesters, 0);
 
     addToRunTimeSelectionTable
     (
-        coalesenceFrequencyKernel,
-        CoulaloglouAndTavlarides,
+        coalescenceEfficiencyKernel,
+        Chesters,
         dictionary
     );
 }
@@ -47,40 +48,55 @@ namespace coalesenceFrequencyKernels
 
 // * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
 
-Foam::coalesenceFrequencyKernels::CoulaloglouAndTavlarides::
-CoulaloglouAndTavlarides
+Foam::coalescenceEfficiencyKernels::Chesters::Chesters
 (
     const dictionary& dict,
     const fvMesh& mesh
 )
 :
-    coalesenceFrequencyKernel(dict, mesh),
+    coalescenceEfficiencyKernel(dict, mesh),
     fluid_(mesh.lookupObject<twoPhaseSystem>("phaseProperties")),
-    epsilon_(fluid_.phase2().turbulence().epsilon()())
+    Ceff_(dict.lookup("Ceff")),
+    ReExp_(dict.lookup("ReExp")),
+    WeExp_(dict.lookup("WeExp"))
 {}
 
 
 // * * * * * * * * * * * * * * * * Destructor  * * * * * * * * * * * * * * * //
 
-Foam::coalesenceFrequencyKernels::CoulaloglouAndTavlarides::
-~CoulaloglouAndTavlarides()
+Foam::coalescenceEfficiencyKernels::Chesters::~Chesters()
 {}
 
 
 // * * * * * * * * * * * * * * * Member Functions  * * * * * * * * * * * * * //
 
 Foam::tmp<Foam::volScalarField>
-Foam::coalesenceFrequencyKernels::CoulaloglouAndTavlarides::omega
+Foam::coalescenceEfficiencyKernels::Chesters::Pc
 (
     const label nodei,
     const label nodej
 ) const
 {
+    const phasePair& pair = fluid_.pair1In2();
     const volScalarField& d1 = fluid_.phase1().ds(nodei);
     const volScalarField& d2 = fluid_.phase1().ds(nodej);
+    volScalarField We(pair.We());
+    volScalarField xi(d1/d2);
+    volScalarField theta
+    (
+        "theta",
+        Ceff_
+       *pow(max(pair.Re(), SMALL), ReExp_)
+       *pow(max(We, SMALL), WeExp_)
+    );
+
     return
-        cbrt(epsilon_)*sqr(d1 + d2)
-       *sqrt(pow(d1, 2.0/3.0) + pow(d2, 2.0/3.0));
+        Foam::exp
+        (
+          - theta*sqrt(We)
+           *sqrt(0.75*(1.0 + sqr(xi))*(1.0 + pow3(xi)))
+           /(sqrt(fluid_.phase1().rho()/fluid_.phase2().rho())*pow3(1.0 + xi))
+        );
 }
 
 // ************************************************************************* //
