+++ applications/solvers/multiphase/interfacialModels/coalescenceKernels/coalescenceEfficiencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.H
@@ -2,7 +2,7 @@
   =========                 |
   \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
    \\    /   O peration     |
-    \\  /    A nd           | Copyright (C) 2017 Alberto Passalacqua
+    \\  /    A nd           | Copyright (C) 2015-2017 Alberto Passalacqua
      \\/     M anipulation  |
 -------------------------------------------------------------------------------
 License
@@ -22,27 +22,38 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::populationBalanceSubModels::aggregationKernels::
+    Foam::populationBalanceSubModels::coalescenceEfficiencyKernels::
     CoulaloglouAndTavlarides
 
 Description
-    Coalesence frequency kernel based on kinetic theory of gasses
+    Coalesence efficiency based on film drainage and rupture times.
 
     \f[
-        \omega = \epsilon_f^{1/3}(d_i^{2/3} + d_j^{2/3})
+        P_c =
+            \text{exp}
+            \left(
+              - C_{eff} \frac{\mu_f \rho_f \epsilon_f}{\sigma^2}
+                \left(
+                    \frac{d_i^{1/3} d_j^{1/3}}{d_i^{1/3} + d_j^{1/3}}
+                \right)
+            \right)
     \f]
 
     where
 
     \vartable
+        \sigma      |  Surface tension
+        \rho_f      |  Density of liquid phase [kg/m3]
         \epsilon_f  |  Turbulent dissipation rate of liquid phase [m2/s3]
+        \mu_f       |  Molecular dynamic iscosity of liquid phase [Pa s]
         d_i         |  Diameter of bubble i [m]
         d_j         |  Diameter of bubble j [m]
     \endvartable
 
     References
     \verbatim
-        "Description of interaction processes in agitated liquid-liquid dis[ersions"
+        "Description of interaction processes in agitated liquid-liquid
+        dis[ersions"
         C Coulaloglou, L Tavlarides
         Chemical Engineering Science
         Volume 32, Pages 1289-1297, 1977
@@ -56,7 +67,7 @@ SourceFiles
 #ifndef CoulaloglouAndTavlarides_H
 #define CoulaloglouAndTavlarides_H
 
-#include "coalesenceFrequencyKernel.H"
+#include "coalescenceEfficiencyKernel.H"
 #include "twoPhaseSystem.H"
 #include "phaseModel.H"
 #include "PhaseCompressibleTurbulenceModel.H"
@@ -65,7 +76,7 @@ SourceFiles
 
 namespace Foam
 {
-namespace coalesenceFrequencyKernels
+namespace coalescenceEfficiencyKernels
 {
 
 /*---------------------------------------------------------------------------*\
@@ -74,15 +85,16 @@ namespace coalesenceFrequencyKernels
 
 class CoulaloglouAndTavlarides
 :
-    public coalesenceFrequencyKernel
+    public coalescenceEfficiencyKernel
 {
     // Private data
 
         //- Refrence to two phase system
         const twoPhaseSystem& fluid_;
 
-        //- Turbulent dissipation rate field
-        const volScalarField& epsilon_;
+        //- Efficiency coefficient
+        dimensionedScalar Ceff_;
+
 
 public:
 
@@ -107,7 +119,7 @@ public:
     // Member Functions
 
         //- Aggregation kernel
-        virtual tmp<volScalarField> omega
+        virtual tmp<volScalarField> Pc
         (
             const label nodei,
             const label nodej
@@ -118,7 +130,7 @@ public:
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
-} // End namespace coalesenceFrequencyKernels
+} // End namespace coalescenceEfficiencyKernels
 } // End namespace Foam
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
