+++ applications/solvers/multiphase/interfacialModels/coalescenceKernels/coalescenceFrequencyKernels/CoulaloglouAndTavlarides/CoulaloglouAndTavlarides.H
@@ -2,7 +2,7 @@
   =========                 |
   \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
    \\    /   O peration     |
-    \\  /    A nd           | Copyright (C) 2015-2017 Alberto Passalacqua
+    \\  /    A nd           | Copyright (C) 2017 Alberto Passalacqua
      \\/     M anipulation  |
 -------------------------------------------------------------------------------
 License
@@ -22,87 +22,78 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::populationBalanceSubModels::coalesenceEfficiencyKernel
+    Foam::populationBalanceSubModels::aggregationKernels::
+    CoulaloglouAndTavlarides
 
 Description
-    Abstract class for coalesence effeciency kernels.
+    Coalesence frequency kernel based on kinetic theory of gasses
+
+    \f[
+        \omega = \epsilon_f^{1/3}(d_i^{2/3} + d_j^{2/3})
+    \f]
+
+    where
+
+    \vartable
+        \epsilon_f  |  Turbulent dissipation rate of liquid phase [m2/s3]
+        d_i         |  Diameter of bubble i [m]
+        d_j         |  Diameter of bubble j [m]
+    \endvartable
+
+    References
+    \verbatim
+        "Description of interaction processes in agitated liquid-liquid dis[ersions"
+        C Coulaloglou, L Tavlarides
+        Chemical Engineering Science
+        Volume 32, Pages 1289-1297, 1977
+    \endverbatim
 
 SourceFiles
-    coalesenceEfficiencyKernel.C
+    CoulaloglouAndTavlarides.C
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef coalesenceEfficiencyKernel_H
-#define coalesenceEfficiencyKernel_H
+#ifndef CoulaloglouAndTavlarides_H
+#define CoulaloglouAndTavlarides_H
 
-#include "dictionary.H"
-#include "volFields.H"
-#include "dimensionedTypes.H"
-#include "runTimeSelectionTables.H"
+#include "coalescenceFrequencyKernel.H"
+#include "twoPhaseSystem.H"
+#include "phaseModel.H"
+#include "PhaseCompressibleTurbulenceModel.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 namespace Foam
 {
+namespace coalescenceFrequencyKernels
+{
 
 /*---------------------------------------------------------------------------*\
-                    Class coalesenceEfficiencyKernel Declaration
+                    Class CoulaloglouAndTavlarides Declaration
 \*---------------------------------------------------------------------------*/
 
-class coalesenceEfficiencyKernel
+class CoulaloglouAndTavlarides
+:
+    public coalescenceFrequencyKernel
 {
-    // Private member functions
+    // Private data
 
-        //- Disallow default bitwise copy construct
-        coalesenceEfficiencyKernel(const coalesenceEfficiencyKernel&);
+        //- Refrence to two phase system
+        const twoPhaseSystem& fluid_;
 
-        //- Disallow default bitwise assignment
-        void operator=(const coalesenceEfficiencyKernel&);
-
-
-protected:
-
-    // Protected data
-
-        //- Dictionary
-        const dictionary& dict_;
-
-        //- Mesh
-        const fvMesh& mesh_;
-
+        //- Turbulent dissipation rate field
+        const volScalarField& epsilon_;
 
 public:
 
-    //- Runtime type information
-    TypeName("coalesenceEfficiencyKernel");
-
-    // Declare runtime constructor selection table
-    declareRunTimeSelectionTable
-    (
-        autoPtr,
-        coalesenceEfficiencyKernel,
-        dictionary,
-        (
-            const dictionary& dict,
-            const fvMesh& mesh
-        ),
-        (dict, mesh)
-    );
+        //- Runtime type information
+        TypeName("CoulaloglouAndTavlarides");
 
 
     // Constructors
 
         //- Construct from components
-        coalesenceEfficiencyKernel
-        (
-            const dictionary& dict,
-            const fvMesh& mesh
-        );
-
-
-    // Selectors
-
-        static autoPtr<coalesenceEfficiencyKernel> New
+        CoulaloglouAndTavlarides
         (
             const dictionary& dict,
             const fvMesh& mesh
@@ -110,22 +101,24 @@ public:
 
 
     //- Destructor
-    virtual ~coalesenceEfficiencyKernel();
+    virtual ~CoulaloglouAndTavlarides();
 
 
     // Member Functions
 
         //- Aggregation kernel
-        virtual tmp<volScalarField> Pc
+        virtual tmp<volScalarField> omega
         (
             const label nodei,
             const label nodej
-        ) const = 0;
+        ) const;
+
 };
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
+} // End namespace coalescenceFrequencyKernels
 } // End namespace Foam
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
