+++ applications/solvers/multiphase/interfacialModels/coalescenceKernels/coalescenceFrequencyKernels/PrinceAndBlanch/PrinceAndBlanch.H
@@ -2,7 +2,7 @@
   =========                 |
   \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
    \\    /   O peration     |
-    \\  /    A nd           | Copyright (C) 2015-2017 Alberto Passalacqua
+    \\  /    A nd           | Copyright (C) 2017 Alberto Passalacqua
      \\/     M anipulation  |
 -------------------------------------------------------------------------------
 License
@@ -22,21 +22,60 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::populationBalanceSubModels::coalesenceEfficiencyKernels::
-    constant
+    Foam::populationBalanceSubModels::aggregationKernels::
+    PrinceAndBlanch
 
 Description
-    Constant coalesence efficiency
+    Coalesence frequency kernel based on turbulence, buoyancy, and
+    laminar shear.
+
+    \f[
+        \omega_{turb} =
+            0.089 \pi (d_i + d_j)^2
+            \sqrt{d_i^{\frac{2}{3}} + d_j^{\frac{2}{3}} }
+            epsilon_f^{frac{1}{3}}
+
+        \omega_{buoyancy} =
+            \pi (d_i + d_j)^2
+            (
+               \sqrt{2.14*\sigma/(d_i \rho_f) + 0.5 \mathbf{g} d_i}
+             - \sqrt{2.14*\sigma/(d_j \rho_f) + 0.5 \mathbf{g} d_j}
+            )
+
+        \omega_{shear} =
+            \frac{2}{3} (d_i + d_j)^3 |\nabla \mathbf{U}_f|
+
+    \f]
+
+    where
+
+    \vartable
+        \sigma      |  Surface tension
+        \mathbf{U}_f|  Liquid velocity [m/s]
+        \rho_f      |  Density of liquid phase [kg/m3]
+        \epsilon_f  |  Turbulent dissipation rate of liquid phase [m2/s3]
+        g           |  Gravitational acceleration [m/s2]
+        \d_i        |  Diameter of bubble i [m]
+        \d_j        |  Diameter of bubble j [m]
+    \endvartable
+
+    References
+    \verbatim
+        "Bubble Coalescence and Break-up n Air-Sparged Bubble Columns"
+        M Prince and Harvey Blanch
+        AIChE Journal
+        Volume 36, Pages 1485 - 1499, 1990
+    \endverbatim
 
 SourceFiles
-    constant.C
+    PrinceAndBlanch.C
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef constant_H
-#define constant_H
+#ifndef PrinceAndBlanch_H
+#define PrinceAndBlanch_H
 
-#include "coalesenceEfficiencyKernel.H"
+#include "coalescenceFrequencyKernel.H"
 #include "twoPhaseSystem.H"
 #include "phaseModel.H"
 #include "PhaseCompressibleTurbulenceModel.H"
@@ -45,36 +84,42 @@ SourceFiles
 
 namespace Foam
 {
-namespace coalesenceEfficiencyKernels
+namespace coalescenceFrequencyKernels
 {
 
 /*---------------------------------------------------------------------------*\
-                    Class constant Declaration
+                    Class PrinceAndBlanch Declaration
 \*---------------------------------------------------------------------------*/
 
-class constant
+class PrinceAndBlanch
 :
-    public coalesenceEfficiencyKernel
+    public coalescenceFrequencyKernel
 {
     // Private data
 
         //- Refrence to two phase system
         const twoPhaseSystem& fluid_;
 
-        //- Efficiency coefficient
-        dimensionedScalar Ceff_;
+        //- Switch for turbulence driven coalescence
+        Switch turbulent_;
+
+        //- Switch for buoyancy drive coalescence
+        Switch buoyant_;
+
+        //- Switch for laminar shear driven coalescence
+        Switch LS_;
 
 
 public:
 
         //- Runtime type information
-        TypeName("constant");
+        TypeName("PrinceAndBlanch");
 
 
     // Constructors
 
         //- Construct from components
-        constant
+        PrinceAndBlanch
         (
             const dictionary& dict,
             const fvMesh& mesh
@@ -82,13 +127,13 @@ public:
 
 
     //- Destructor
-    virtual ~constant();
+    virtual ~PrinceAndBlanch();
 
 
     // Member Functions
 
         //- Aggregation kernel
-        virtual tmp<volScalarField> Pc
+        virtual tmp<volScalarField> omega
         (
             const label nodei,
             const label nodej
@@ -99,7 +144,7 @@ public:
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
-} // End namespace coalesenceEfficiencyKernels
+} // End namespace coalescenceFrequencyKernels
 } // End namespace Foam
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
